Phase 2-3 - Implementation.

+++ AritySystems/Models/PriceUpdateModel.cs
@@ -8,7 +8,7 @@ namespace AritySystems.Models
     public class PriceUpdateModel
     {
         public int ItemId { get; set; }
-        public int DollerPrice { get; set; }
-        public int RMBPrice { get; set; }
+        public decimal DollerPrice { get; set; }
+        public decimal RMBPrice { get; set; }
     }
 }
